small changes
learned how to use stack.protected to switch views by user case; login or not. implemented simple buttons for register and sign-in. register takes you to another page, and login just takes to tabs. trying to implement x button on profile tab to take you back to sign-in screen, in progress sign-out button